+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Drop package.json functions https://github.com/archetype-themes/plugin-theme-component/pull/299
+
 ## [1.9.7] - 2024-01-17
 
 ### Changed